# Performance improvement PersonaBar tree view node selection change