@nadeem-akhtar-nu 2020-02-02
> @nadeem-akhtar-nu this seems to be a duplicate of #3533

@david-poindexter 
This is a different method change with different use case. It is a similar change to #3533, though.